+++ src/Component/OpenApi3/Generator/Endpoint/GetGetUriTrait.php
@@ -6,6 +6,7 @@ use Jane\Component\JsonSchemaRuntime\Reference;
 use Jane\Component\OpenApi3\Generator\EndpointGenerator;
 use Jane\Component\OpenApi3\Guesser\GuessClass;
 use Jane\Component\OpenApi3\JsonSchema\Model\Parameter;
+use Jane\Component\OpenApi3\JsonSchema\Model\Schema;
 use Jane\Component\OpenApiCommon\Guesser\Guess\OperationGuess;
 use PhpParser\Node\Arg;
 use PhpParser\Node\Expr;
@@ -17,17 +18,24 @@ trait GetGetUriTrait
 {
     public function getGetUri(OperationGuess $operation, GuessClass $guessClass): Stmt\ClassMethod
     {
-        $names = [];
+        $names = $types = [];
 
         foreach ($operation->getParameters() as $parameter) {
             if ($parameter instanceof Reference) {
                 $parameter = $guessClass->resolveParameter($parameter);
             }
 
-            if ($parameter instanceof Parameter && EndpointGenerator::IN_PATH === $parameter->getIn()) {
-                // $url = str_replace('{param}', $param, $url)
-                $names[] = $parameter->getName();
+            if (!$parameter instanceof Parameter || EndpointGenerator::IN_PATH !== $parameter->getIn()) {
+                continue;
             }
+
+            $schema = $parameter->getSchema();
+            if ($schema instanceof Reference) {
+                [, $schema] = $guessClass->resolve($parameter->getSchema(), Schema::class);
+            }
+
+            $names[] = $parameter->getName();
+            $types[] = $schema instanceof Schema ? $schema->getType() : null;
         }
 
         if (\count($names) === 0) {
@@ -47,9 +55,13 @@ trait GetGetUriTrait
                     new Arg(new Expr\Array_(array_map(function ($name) {
                         return new Scalar\String_('{' . $name . '}');
                     }, $names))),
-                    new Arg(new Expr\Array_(array_map(function ($name) {
-                        return new Expr\PropertyFetch(new Expr\Variable('this'), $name);
-                    }, $names))),
+                    new Arg(new Expr\Array_(array_map(function ($type, $name) {
+                        return 'array' === $type
+                            // return str_replace(['{param}'], [implode(',', $this->param)], '/path/{param}')
+                            ? new Expr\FuncCall(new Name('implode'), [new Arg(new Scalar\String_(',')), new Arg(new Expr\PropertyFetch(new Expr\Variable('this'), $name))])
+                            // return str_replace(['{param}'], [$this->param], '/path/{param}')
+                            : new Expr\PropertyFetch(new Expr\Variable('this'), $name);
+                    }, $types, $names))),
                     new Arg(new Scalar\String_($operation->getPath())),
                 ])),
             ],